refactor(vscode): split lsp releated code to `linter.ts` (#15789)
> This PR refactors the VSCode extension by extracting LSP-related functionality from extension.ts into a new linter.ts module, improving code organization and separation of concerns.
>
> Key Changes:
>
> - Moved LSP client initialization, configuration, and lifecycle management from extension.ts to linter.ts
> -  Created a new commands.ts file to centralize command constants
> -  Simplified extension.ts to focus on extension-level orchestration by delegating to linter-specific functions
