+++ editors/vscode/client/commands.ts
@@ -1,0 +1,9 @@
+const commandPrefix = 'oxc';
+
+export const enum OxcCommands {
+  RestartServer = `${commandPrefix}.restartServer`,
+  ShowOutputChannel = `${commandPrefix}.showOutputChannel`,
+  ToggleEnable = `${commandPrefix}.toggleEnable`,
+  // only for linter.ts usage
+  ApplyAllFixesFile = `${commandPrefix}.applyAllFixesFile`,
+}
